// src/types/bonusMalus.ts
export interface BonusMalusCalculation {
  team1Chutes: number;
  team2Chutes: number;
  team1EpicerieAlarms: number;
  team2EpicerieAlarms: number;
  team1VousEtesNulsCount: number;
  team2VousEtesNulsCount: number;
  team1BonusPoints: number;
  team2BonusPoints: number;
  team1MalusPoints: number;
  team2MalusPoints: number;
  team1FinalPoints: number;
  team2FinalPoints: number;
}